changed chars to ints

+++ src/framebuffer.h
@@ -7,8 +7,8 @@
 
 class Framebuffer {
 private:
-    uint8_t fbwidth;
-    uint8_t fbheight;
+    uint16_t fbwidth;
+    uint16_t fbheight;
 
 	std::vector<uint8_t> fb;
     std::string buffer;
@@ -16,11 +16,11 @@ private:
 
 public:
 
-uint8_t GetWidth();
+uint16_t GetWidth();
 
-uint8_t GetHeight();
+uint16_t GetHeight();
 
-void PutPixel(uint8_t x, uint8_t y, bool state);
+void PutPixel(uint16_t x, uint16_t y, bool state);
 
 // swap bits a and b in a byte
 uint8_t SwapBits(uint8_t byte, unsigned int a, unsigned int b);
@@ -32,7 +32,7 @@ uint8_t ByteToBraille(uint8_t byte);
 // Translates a codepoint to a Unicode character.
 std::string UnicodeToUTF8(unsigned int codepoint);
 
-uint8_t GetByte(uint8_t x, uint8_t y);
+uint8_t GetByte(uint16_t x, uint16_t y);
 
 std::string BitmapToBraille();
 
@@ -40,7 +40,7 @@ void BufferText(uint16_t x, uint16_t y, std::string msg);
 	
 void Refresh();
 
-Framebuffer(uint8_t x, uint8_t y);
+Framebuffer(uint16_t x, uint16_t y);
 };
 
 #endif
